Migrate module setup to script plugin
